WIP: patch kernel to skip endbr32 instruction on real i486.
This is stupid patch which just hacks invalid instruction
operand trap code to skip endbr32 instruction.

endbr32 is part of Intel’s Control-flow Enforcement Technology and
older processors should handle this instruction (4 bytes) as a nop.
However, it's true for Pentium Pro and higher. But i486s and original
Pentiums are not able to handle this as a multi nop and just ends in
invalid instruction operand trap, which sends SIGILL to process.

GCC, even when -fcf-protection=none is used, when GCC itself is
compiled without --disable-cet, have build-in functions "polluted"
with endbr32 (for example: some floating point functions like:
__mulxc3, __moddi3, __divdi3, __divmoddi4, __udivdi3) which then
later gets into musl libc library so busybox going to die in pain.

There is nice and clean solution: compile GCC without CET support
but I'm not able to do it now (cos I'm stupid).

PS: qemu with cpu set to i486 just works, so it's not
true emulation of i486 CPU.